+++ tests/unit/test_power_transformer.py
@@ -5,7 +5,7 @@ import xarray as xr
 from sklearn.preprocessing import PowerTransformer
 
 import mesmer
-from mesmer.core.utils import _check_dataarray_form, _check_dataset_form
+from mesmer.core.utils import _check_dataarray_form
 from mesmer.stats._power_transformer import (
     _yeo_johnson_inverse_transform_np,
     _yeo_johnson_optimize_lambda_np,
@@ -234,15 +234,15 @@ def test_power_transformer_xr():
 
     # 1 - fitting
     pt_coefficients = mesmer.stats.fit_yeo_johnson_transform(
-        monthly_residuals, yearly_T
+        yearly_T, monthly_residuals
     )
     # 2 - transformation
     transformed = mesmer.stats.yeo_johnson_transform(
-        monthly_residuals, pt_coefficients, yearly_T
+        yearly_T, monthly_residuals, pt_coefficients
     )
     # 3 - back-transformation
     inverse_transformed = mesmer.stats.inverse_yeo_johnson_transform(
-        transformed.transformed, pt_coefficients, yearly_T
+        yearly_T, transformed.transformed, pt_coefficients
     )
 
     xr.testing.assert_allclose(
@@ -274,15 +274,12 @@ def test_power_transformer_xr():
         required_dims=("cells", "time"),
         shape=(n_gridcells, n_years * 12),
     )
-    _check_dataset_form(
-        pt_coefficients, name="pt_coefficients", required_vars=("lambda_coeffs")
-    )
     _check_dataarray_form(
-        pt_coefficients.lambda_coeffs,
+        pt_coefficients,
         name="lambda_coeffs",
         ndim=3,
         required_dims=("cells", "coeff", "month"),
         shape=(12, n_gridcells, 2),
     )
-    assert "month" in pt_coefficients.lambda_coeffs.coords
+    assert "month" in pt_coefficients.coords
     xr.testing.assert_equal(expected_month, pt_coefficients.month)